@@ -1,6 +1,14 @@
+import 'package:bytebank/screens/contacts/form.dart';
+import 'package:bytebank/screens/contacts/list.dart';
+import 'package:bytebank/screens/dashboard.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(ByteBankApp());
+import 'database/app_database/app_datebase.dart';
+
+void main() {
+  runApp(ByteBankApp());
+  createDatabase();
+}
 
 class ByteBankApp extends StatelessWidget {
   @override
@@ -15,56 +23,5 @@ class ByteBankApp extends StatelessWidget {
           ),
         ),
         home: Dashboard());
-  }
-}
-
-class Dashboard extends StatelessWidget {
-  const Dashboard({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Dashboard'),
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Image.asset('images/bytebank_logo.png')),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Container(
-              padding: const EdgeInsets.all(8.0),
-              height: 100,
-              width: 150,
-              color: Theme.of(context).primaryColor,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Icon(
-                    Icons.people,
-                    color: Colors.white,
-                    size: 24.0,
-                  ),
-                  Text(
-                    'Contacts',
-                    style: TextStyle(
-                      color: Colors.white,
-                      fontSize: 16.0,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          )
-        ],
-      ),
-    );
   }
 }
